+++ Vistas/AgregarProductos.cs
@@ -70,7 +70,7 @@ namespace Tienda_de_Ropa
             
                 if (inputsValidos)
                 {
-                    Producto producto = new Producto(int.Parse(txt_id.Text), txt_nombre.Text, txt_descripcion.Text, double.Parse(txt_precio.Text));
+                    Producto producto = new Producto(int.Parse(txt_id.Text), txt_nombre.Text, txt_descripcion.Text, decimal.Parse(txt_precio.Text));
                     Producto_Controlador.crearProducto(producto);
                     MessageBox.Show("Producto creado con exito");
                     AgregarProductos form1 = new AgregarProductos();
